requirements.txt
gmusicapi==12.1.1
python-vlc>=3.0.7110
pynput==1.4.5
